Implement device_filter in inventory_to_container
Implement a methodology to test if device is part of device_filter or
not.

Will be used as a contraint to define if container should be added to
CVP configuration file.

With a device_filter=['DC1-LE'], Container topology becomes (We can see only devices matching filter):

```
CVP_TOPOLOGY:
  DC1_FABRIC:
    parent_container: Tenant
  DC1_L2LEAF1:
    devices: []
    parent_container: DC1_L2LEAFS
  DC1_L2LEAF2:
    devices: []
    parent_container: DC1_L2LEAFS
  DC1_L2LEAFS:
    parent_container: DC1_FABRIC
  DC1_L3LEAFS:
    parent_container: DC1_FABRIC
  DC1_LEAF1:
    devices:
    - DC1-LEAF1A
    - DC1-LEAF1B
    parent_container: DC1_L3LEAFS
  DC1_LEAF2:
    devices:
    - DC1-LEAF2A
    - DC1-LEAF2B
    parent_container: DC1_L3LEAFS
  DC1_SPINES:
    devices: []
    parent_container: DC1_FABRIC
changed: false
```