Add support for un-provisioned device in CV #407
When you want to provision a complete topoplogy and devices are not
already in Cloudvision, you can configure inventory to ignore these
devices by using a host variable: `is_deployed`

- `is_deployed: true` or `is_deplpyed is not defined`: An entry in
  __cv_device__ is generated and AVD will configure device on
Cloudvision. If device is undefined, an error is raised.

- `is_deployed: false`: Device is not configured in __cv_device__
  topology and only its configlet is uploaded on Cloudvision.

Here is an overview with the key configured in the YAML inventory:

```yaml
  DC1_BL1:
    hosts:
      DC1-BL1A:
        ansible_port: 8012
  DC1_BL2:
    hosts:
      DC1-BL2A:
        ansible_port: 8012
        # Device configuration is generated by AVD
        # Device is not configured on Cloudvision (configlet is
uploaded)
        is_deployed: false
```