Requirements update + relocation (#824)
* Add a requirements file
This PR is a follow up to an email sent to the maintainers of this certified collection notifying them of a few new changes that will be needed for the next Ansible Automation Platform release. Contained in this PR is an example python requirements file built specifically for your collection. You can elect to merge this as is and edit as needed or close the PR and use it as an example to base your own commit off of. if you have any questions about why this PR was sent or confused about its necessity, please don't hesitate to ask in the comments here or email us at ansiblepartners@redhat.com

* Move requirements to collection root + Req update

* fix PR template document

* Update Release notes about Python requirements

* dep(cvprac): Set cvprac to v1.0.5

Co-authored-by: Anshul Behl <anshulbehl.19@gmail.com>