Doc(eos_designs): Marked switch and host ports as required
connected_endpoint.adapters.endpoint_ports is required for generating the
network documentation, and the switch and switchport assignments don't
cause an error if neither is assign, but conceptually endpoints should
only be defined if they're connected to the fabric.

Possible future work is to make connected_endpoint.adapters.endpoint_ports
optional since the concept of endpoint interface assignments may be
not applicable in some network applications.