Fix(eos_cli_config_gen): Block delimers between Extended IPv4 ACLs
It looks like when the exclamation points were refactored in
be813b3381242aefb3620c1b131e361c556aa2ef the new ! was placed one line
off and fell outside the for loop, so only a single ! was generated for all the
IPv4 ACLs.

Also aligning Extended IPv4 ACLs with other blocks to not use an if defined block
and rely on looping over an empty list. Fixing indents accordingly.