Fix(eos_designs): Pull endpoint_ports from combined settings
endpoint_ports was incorrectly being pulled from the original raw adapter
configuration, which did not allow the endpoint port to be configured in
a port profile, which would be useful when stamping out a large number of
connected endpoints with the same port configuraiton.

Also makes it much easier to stub out endpoint_ports when they're not
applicable for theuser since they can set endpoint_ports: [ 'N/A' ] in
just a parent port profile template.