Fix(eos_config_deploy_cvp): fix pylint warnings
AnsibleVaultEncryptedUnicode belongs to ansible.parsing.yaml which sadly
cannot be imported in a custom module (as only ansible.module_utils is
allowed)

The temporary solution is to replace the constructor with a lambda
function that import the `!vault` value as `!VAULT`.

This is a temporary solution as there is a plan to replace the import of
the inventory (which supports only YAML format) with a more agnostic
approach using internal ansible variables.