Issue 3760 - Add knob for IS-IS system-id format
This commit adds support for the IS-IS system-id to be based on the
underlay loopback interface IP via the gloabl `isis_system_id_format`
key. For now AVD will still use the default behavior of the `node id`
field to generate the IS-IS system-id. In AVD 5.0 the default behavior
is planned to be moved to the `underlay_loopback` format.

Includes all tests for both knob settings. No existing molecule tests
were impacted, validating the default behavior works as expected.